submodule.c: store new submodule commits in a struct
In order to teach `git fetch --recurse-submodules` how to fetch changed
submodules that are not at HEAD, we need to know the commit id and path
where the submodule was found. We can store this information in the
string_list of changed submodule names.

Begin this refactor by replacing the changed submodules string_list's
string_list_item.util (which is currently struct oid_array *) with a new
struct. This struct does not hold any new data yet; that will be added
in subsequent commits.

Signed-off-by: Glen Choo <chooglen@google.com>