cover letter: teach fetch --recurse-submodules to handle submodules not in the index
After a fetch in the superproject, fetch_populated_submodules() checks
for new, missing, submodule commits and fetches them. This ensures that
`git checkout --recurse-submodules newly-fetched-branch` can succeed
because the submodule already has the appropriate commit.

The current implementation isn't foolproof because, as noted in
Documentation/git-fetch.txt, only submodules in the index are fetched,
thus this workflow fails if a newly fetched commit contains a submodule
that is not in the index. This can happen in a few cases:

a) a completely new submodule was added in the new commit
b) the new commit references a submodule that the repo has already
   initialized, but isn't in the index

This suggests that we want to teach submodule.c to handle submodules
that are not in the index. Git already knows how to read submodules from
trees e.g. submodule_from_path() takes a treeish_name parameter.