upload-pack: use lookup fn for pack objects hook
Implement git_protected_config_get_string(), which gets a
string from protected_config, and use it to get the value of
`uploadpack.packObjectsHook`. This avoids iterating through the whole
protected config and removes an unnecessary const declaration from
upload_pack_data.pack_objects_hook (it was declared const to match the
return type of git_config_string(), but we have to free() it anyway).

Signed-off-by: Glen Choo <chooglen@google.com>