setup: use lookup function for safe.directory
Implement git_protected_config_get_value_multi(), which gets a
multivalued string from protected_config, and use it to get the value of
`safe.directory`.

Since we no longer iterate the config with proteted_config(),
safe_directory_data is removed (it was only needed to preserve the state
of the config iteration).

Signed-off-by: Glen Choo <chooglen@google.com>