CodingGuidelines: recommend 0-initializing with {0}
Our C codebase uses both "memset(&ptr, 0, sizeof(ptr))" and
"= { 0 }" to initialize auto variables. For consistency, recommend the
latter as the only approach.

Besides being more terse, it is more correct because the language
initializes the missing members in the same way as static variables (see
C99 6.7.8.21), which ensures that the values are reasonable; the
memset() approach assumes that variables should be initialized to all
0s, which may not be true on certain platforms (though this is
admittedly rare).

Signed-off-by: Glen Choo <chooglen@google.com>