fetch: refactor --submodule-prefix
Since "git fetch" learned to recurse into submodules [1], it has passed
"--submodule-prefix=<path from root of superproject tree to submodule>" to print
full paths to the submodules being fetched. This is the same value as
"--super-prefix"'s when calling "git read-tree --recurse-submodules", as well as
"grep" and "ls-files" (before they were implemented in-process [2] [3]).

Let's standardize the way we pass such prefixes. Having a per-command
"--submodule-prefix" is preferable, since it gives us fine-grained control over
which commands support it and which don't, and we can preserve the effort-saving
properties of the global flag by sharing an implementation under the hood.

Refactor "git fetch" so that "--submodule-prefix" is stored in
repository.submodule_prefix, and create functions in submodule.h for parsing and
reading its value. Also, mark it with PARSE_OPT_HIDDEN and drop it from our
documentation. In a future commit, we will follow this pattern for "git
read-tree".

Since we've previously documented that repository.submodule_prefix is only set
for submodules, introduce repository.is_submodule for distinguishing between
the_repository and a submodule (though this turns out to be unused, since no
caller has ever used repository.submodule_prefix for that purpose).

[1] 7dce19d374 (fetch/pull: Add the --recurse-submodules option, 2010-11-12)
[2] 188dce131f (ls-files: use repository object, 2017-06-22)
[3] f9ee2fcdfa (grep: recurse in-process using 'struct repository', 2017-08-02)