config: plumb config_source through static fns
This reduces the direct dependence on the "cf" static variable, which
will make it easier to remove in a later commit. The plumbed arg is
named "cs" to differentiate between it and the static variable.

The remaining references to "cf" will be tackled in later commits, since
they fall into the one following categories that need more work to
remove:

- Functions that assign to "cf"
- Public functions
- Callback functions