config: create lib, move create_*_config_sources()
Create lib/config.c, which will contain config library code. The first
bit of functionality we'll add to it is to read config files/buffers
without interfering with global state.

That's a lot of code though, so for starters, we'll move the config
source-creating logic (leaving the config parsing machinery untouched).
Its inclusion in lib/config.h is only temporary; we can remove it once
we move all of the config file/buffer-reading functionality.