config.c: create config_source lifecycle fns
To make the "struct config_source" initialization and release logic
easier to follow, refactor it out from the various functions they're
implemented in, and rearrange them so that initialization of all the
fields happens together. This is okay to do, since the source-specific
fields (e.g. "file" and "path" for config files, "buf" and "len" for
buffers) are disjoint from the general fields (e.g. "value", "var")
anyway.