(RFC-only) config: combine config_fn_t args
Change the config_fn_t signature by combining the "key" and "value"
args, and a new .kvi member that holds the key_value_info, and refactor
the config.c machinery accordingly. None of the implementations are
changed, however, so this commit does not build - I've split this out
for readability, but post-RFC I will squash this with the rest of the
refactor + cocci changes.

Signed-off-by: Glen Choo <chooglen@google.com>