(RFC-only) config: apply cocci to config_fn_t implementations
Change the signature of *most* functions that are invoked as
"config_fn_t" callbacks by applying
contrib/coccinelle/config_fn_kvi.pending.cocci. To minimize the churn in
the function bodies, assign the config_context .key and .value members
(which replace the first two args) to variables with the same names as
the replaced args.

When calling an inner function with the config name and value, only pass
"config_context" as a complete struct if if the inner function is
actually invoked elsewhere as a config callback; it does not matter
whether the function happens have the same signature as config_fn_t.

This commit does not build, but I've split this out for the RFC so that
it's more obvious which changes are automatic vs manual. Post-RFC I will
squash this with the rest of the refactor + cocci changes.

Signed-off-by: Glen Choo <chooglen@google.com>